Support exported {interface,class,function,variable} declarations.
These already generated appropriate externs symbols, but did not
generate a symbol on the exports object of the module, leading to
Closure Compiler failures when importing the symbol.

This code interprets exported deVclared symbols as global symbols that
are accessible under the same name, consistent with previous code.
That's somewhat odd though - we could equally well interpret it as
symbols that the user promises should magically appear on the exports
object. This does not really matter for Closure Compiler compatibility
though.